+++ src/main/scala/com/gu/autoCancel/AutoCancel.scala
@@ -3,7 +3,7 @@ package com.gu.autoCancel
 import com.gu.util.Logging
 import com.gu.util.reader.Types.FailableOp
 import com.gu.util.zuora.ZuoraModels.SubscriptionId
-import com.gu.util.zuora.{ Zuora, ZuoraDeps }
+import com.gu.util.zuora._
 import org.joda.time.LocalDate
 
 object AutoCancel extends Logging {
@@ -14,9 +14,9 @@ object AutoCancel extends Logging {
     val AutoCancelRequest(accountId, subToCancel, cancellationDate) = acRequest
     logger.info(s"Attempting to perform auto-cancellation on account: $accountId")
     val zuoraOp = for {
-      _ <- Zuora.updateCancellationReason(subToCancel).withLogging("updateCancellationReason")
-      _ <- Zuora.cancelSubscription(subToCancel, cancellationDate).withLogging("cancelSubscription")
-      _ <- Zuora.disableAutoPay(accountId).withLogging("disableAutoPay")
+      _ <- ZuoraUpdateCancellationReason(subToCancel).withLogging("updateCancellationReason")
+      _ <- ZuoraCancelSubscription(subToCancel, cancellationDate).withLogging("cancelSubscription")
+      _ <- ZuoraDisableAutoPay(accountId).withLogging("disableAutoPay")
     } yield ()
     zuoraOp.run.run(zuoraDeps)
   }
