+++ src/main/scala/com/gu/autoCancel/AutoCancelDataCollectionFilter.scala
@@ -4,8 +4,9 @@ import com.github.nscala_time.time.OrderingImplicits._
 import com.gu.autoCancel.AutoCancel.AutoCancelRequest
 import com.gu.util.apigateway.ApiGatewayResponse.noActionRequired
 import com.gu.util.reader.Types.{ FailableOp, _ }
-import com.gu.util.zuora.ZuoraModels.{ AccountSummary, Invoice, SubscriptionId }
-import com.gu.util.zuora.{ Zuora, ZuoraDeps }
+import com.gu.util.zuora.ZuoraGetAccountSummary.{ AccountSummary, Invoice }
+import com.gu.util.zuora.ZuoraModels.SubscriptionId
+import com.gu.util.zuora.{ ZuoraDeps, ZuoraGetAccountSummary }
 import com.gu.util.{ Logging, ZuoraRestConfig }
 import okhttp3.{ Request, Response }
 import org.joda.time.LocalDate
@@ -25,7 +26,7 @@ object AutoCancelDataCollectionFilter extends Logging {
     def default(now: LocalDate, response: Request => Response, config: ZuoraRestConfig): ACFilterDeps = {
       ACFilterDeps(
         now,
-        Zuora.getAccountSummary,
+        ZuoraGetAccountSummary.apply,
         response,
         config
       )
