+++ src/main/scala/com/gu/paymentFailure/GetPaymentData.scala
@@ -3,7 +3,7 @@ package com.gu.paymentFailure
 import com.gu.util.Logging
 import com.gu.util.apigateway.ApiGatewayResponse.internalServerError
 import com.gu.util.reader.Types.FailableOp
-import com.gu.util.zuora.ZuoraModels.{ InvoiceItem, InvoiceTransactionSummary, ItemisedInvoice }
+import com.gu.util.zuora.ZuoraGetInvoiceTransactions.{ InvoiceItem, InvoiceTransactionSummary, ItemisedInvoice }
 import org.joda.time.LocalDate
 
 import scalaz.\/
