+++ src/main/scala/com/gu/util/apigateway/ApiGatewayRequest.scala
@@ -1,7 +1,7 @@
 package com.gu.util.apigateway
 
-import play.api.libs.json._
 import play.api.libs.functional.syntax._
+import play.api.libs.json._
 
 case class RequestAuth(apiToken: String)
