+++ src/main/scala/com/gu/util/zuora/Zuora.scala
@@ -1,169 +1,138 @@
 package com.gu.util.zuora
 
-import com.gu.stripeCustomerSourceUpdated.{ StripeCustomerId, _ }
 import com.gu.util.ZuoraRestConfig
-import com.gu.util.apigateway.ApiGatewayResponse
 import com.gu.util.reader.Types.WithDepsFailableOp
-import com.gu.util.zuora.Zuora.Query
 import com.gu.util.zuora.ZuoraModels._
 import com.gu.util.zuora.ZuoraQueryPaymentMethod.{ AccountId, PaymentMethodId }
 import com.gu.util.zuora.ZuoraReaders._
-import com.gu.util.zuora.ZuoraWriters._
 import com.gu.util.zuora.ZuoraRestRequestMaker._
 import okhttp3.{ Request, Response }
 import org.joda.time.LocalDate
 import play.api.libs.functional.syntax._
-import play.api.libs.json._
-
-import scalaz.{ -\/, NonEmptyList, \/-, _ }
+import play.api.libs.json.{ JsPath, Json, Reads, Writes }
 
 case class ZuoraDeps(response: Request => Response, config: ZuoraRestConfig)
 
-object Zuora {
+object ZuoraGetAccountSummary {
 
-  def getAccountSummary(accountId: String): WithDepsFailableOp[ZuoraDeps, AccountSummary] =
+  case class BasicAccountInfo(id: AccountId, balance: Double, defaultPaymentMethod: PaymentMethodId)
+
+  case class SubscriptionSummary(id: SubscriptionId, subscriptionNumber: String, status: String)
+
+  case class Invoice(id: String, dueDate: LocalDate, balance: Double, status: String)
+
+  case class AccountSummary(basicInfo: BasicAccountInfo, subscriptions: List[SubscriptionSummary], invoices: List[Invoice])
+
+  implicit val basicAccountInfoReads: Reads[BasicAccountInfo] = (
+    (JsPath \ "id").read[String].map(AccountId.apply) and
+    (JsPath \ "balance").read[Double] and
+    (JsPath \ "defaultPaymentMethod" \ "id").read[PaymentMethodId]
+  )(BasicAccountInfo.apply _)
+
+  implicit val invoiceReads: Reads[Invoice] = (
+    (JsPath \ "id").read[String] and
+    (JsPath \ "dueDate").read[LocalDate] and
+    (JsPath \ "balance").read[Double] and
+    (JsPath \ "status").read[String]
+  )(Invoice.apply _)
+
+  implicit val subscriptionSummaryReads: Reads[SubscriptionSummary] = (
+    (JsPath \ "id").read[String].map(SubscriptionId.apply) and
+    (JsPath \ "subscriptionNumber").read[String] and
+    (JsPath \ "status").read[String]
+  )(SubscriptionSummary.apply _)
+
+  implicit val accountSummaryReads: Reads[AccountSummary] = (
+    (JsPath \ "basicInfo").read[BasicAccountInfo] and
+    (JsPath \ "subscriptions").read[List[SubscriptionSummary]] and
+    (JsPath \ "invoices").read[List[Invoice]]
+  )(AccountSummary.apply _)
+
+  def apply(accountId: String): WithDepsFailableOp[ZuoraDeps, AccountSummary] =
     get[AccountSummary](s"accounts/$accountId/summary")
 
-  def getInvoiceTransactions(accountId: String): WithDepsFailableOp[ZuoraDeps, InvoiceTransactionSummary] =
+}
+
+object ZuoraGetInvoiceTransactions {
+  case class InvoiceItem(id: String, subscriptionName: String, serviceStartDate: LocalDate, serviceEndDate: LocalDate, chargeAmount: Double, chargeName: String, productName: String)
+
+  case class ItemisedInvoice(id: String, invoiceDate: LocalDate, amount: Double, balance: Double, status: String, invoiceItems: List[InvoiceItem])
+
+  case class InvoiceTransactionSummary(invoices: List[ItemisedInvoice])
+
+  implicit val invoiceItemReads: Reads[InvoiceItem] = (
+    (JsPath \ "id").read[String] and
+    (JsPath \ "subscriptionName").read[String] and
+    (JsPath \ "serviceStartDate").read[LocalDate] and
+    (JsPath \ "serviceEndDate").read[LocalDate] and
+    (JsPath \ "chargeAmount").read[Double] and
+    (JsPath \ "chargeName").read[String] and
+    (JsPath \ "productName").read[String]
+  )(InvoiceItem.apply _)
+
+  implicit val itemisedInvoiceReads: Reads[ItemisedInvoice] = (
+    (JsPath \ "id").read[String] and
+    (JsPath \ "invoiceDate").read[LocalDate] and
+    (JsPath \ "amount").read[Double] and
+    (JsPath \ "balance").read[Double] and
+    (JsPath \ "status").read[String] and
+    (JsPath \ "invoiceItems").read[List[InvoiceItem]]
+  )(ItemisedInvoice.apply _)
+
+  implicit val invoiceTransactionSummaryReads: Reads[InvoiceTransactionSummary] =
+    (JsPath \ "invoices").read[List[ItemisedInvoice]].map {
+      invoices => InvoiceTransactionSummary(invoices)
+    }
+
+  def apply(accountId: String): WithDepsFailableOp[ZuoraDeps, InvoiceTransactionSummary] =
     get[InvoiceTransactionSummary](s"transactions/invoices/accounts/$accountId")
 
-  def cancelSubscription(subscription: SubscriptionId, cancellationDate: LocalDate): WithDepsFailableOp[ZuoraDeps, Unit] =
+}
+
+object ZuoraCancelSubscription {
+
+  case class SubscriptionCancellation(cancellationEffectiveDate: LocalDate)
+
+  implicit val subscriptionCancellationWrites = new Writes[SubscriptionCancellation] {
+    def writes(subscriptionCancellation: SubscriptionCancellation) = Json.obj(
+      "cancellationEffectiveDate" -> subscriptionCancellation.cancellationEffectiveDate,
+      "cancellationPolicy" -> "SpecificDate",
+      "invoiceCollect" -> false
+    )
+  }
+
+  def apply(subscription: SubscriptionId, cancellationDate: LocalDate): WithDepsFailableOp[ZuoraDeps, Unit] =
     put(SubscriptionCancellation(cancellationDate), s"subscriptions/${subscription.id}/cancel")
 
-  def updateCancellationReason(subscription: SubscriptionId): WithDepsFailableOp[ZuoraDeps, Unit] =
+}
+
+object ZuoraUpdateCancellationReason {
+
+  case class SubscriptionUpdate(cancellationReason: String)
+
+  implicit val subscriptionUpdateWrites = new Writes[SubscriptionUpdate] {
+    def writes(subscriptionUpdate: SubscriptionUpdate) = Json.obj(
+      "CancellationReason__c" -> subscriptionUpdate.cancellationReason
+    )
+  }
+
+  def apply(subscription: SubscriptionId): WithDepsFailableOp[ZuoraDeps, Unit] =
     put(SubscriptionUpdate("System AutoCancel"), s"subscriptions/${subscription.id}")
 
-  def disableAutoPay(accountId: String): WithDepsFailableOp[ZuoraDeps, Unit] =
+}
+
+object ZuoraDisableAutoPay {
+
+  case class AccountUpdate(autoPay: Boolean)
+
+  implicit val accountUpdateWrites = new Writes[AccountUpdate] {
+    def writes(accountUpdate: AccountUpdate) = Json.obj(
+      "autoPay" -> accountUpdate.autoPay
+    )
+  }
+
+  def apply(accountId: String): WithDepsFailableOp[ZuoraDeps, Unit] =
     put(AccountUpdate(autoPay = false), s"accounts/$accountId")
 
-  case class Query(queryString: String)
-
-  case class QueryLocator(value: String) extends AnyVal
-
-  case class QueryResult[QUERYRECORD](records: List[QUERYRECORD], size: Int, done: Boolean, queryLocator: Option[QueryLocator])
-
-  implicit val queryW: Writes[Query] = Json.writes[Query]
-
-  implicit val queryLocator: Format[QueryLocator] =
-    Format[QueryLocator](JsPath.read[String].map(QueryLocator.apply), Writes { (o: QueryLocator) => JsString(o.value) })
-
-  implicit def queryResultR[QUERYRECORD: Reads]: Reads[QueryResult[QUERYRECORD]] = (
-    (JsPath \ "records").read[List[QUERYRECORD]] and
-    (JsPath \ "size").read[Int] and
-    (JsPath \ "done").read[Boolean] and
-    (JsPath \ "queryLocator").readNullable[QueryLocator]
-  ).apply(QueryResult.apply[QUERYRECORD] _)
-
-  case class QueryMoreReq(queryLocator: QueryLocator)
-
-  implicit val wQueryMoreReq: Writes[QueryMoreReq] = Json.writes[QueryMoreReq]
-
-  // https://www.zuora.com/developer/api-reference/#operation/Action_POSTquery
-  def query[QUERYRECORD: Reads](query: Query): WithDepsFailableOp[ZuoraDeps, QueryResult[QUERYRECORD]] =
-    post(query, s"action/query")
-
 }
 
-object SetDefaultPaymentMethod {
-
-  case class SetDefaultPaymentMethod(paymentMethodId: PaymentMethodId)
-
-  implicit val writes = new Writes[SetDefaultPaymentMethod] {
-    def writes(subscriptionUpdate: SetDefaultPaymentMethod) = Json.obj(
-      "DefaultPaymentMethodId" -> subscriptionUpdate.paymentMethodId
-    )
-  }
-
-  def setDefaultPaymentMethod(accountId: AccountId, paymentMethodId: PaymentMethodId): WithDepsFailableOp[ZuoraDeps, Unit] =
-    put(SetDefaultPaymentMethod(paymentMethodId), s"object/account/${accountId.value}")
-
-}
-
-object CreatePaymentMethod {
-
-  case class CreateStripePaymentMethod(
-    accountId: AccountId,
-    cardId: StripeSourceId,
-    customerId: StripeCustomerId,
-    cardCountry: StripeCountry,
-    last4: StripeLast4,
-    expiration: StripeExpiry,
-    creditCardType: CreditCardType
-  )
-
-  sealed abstract class CreditCardType(val value: String)
-  object CreditCardType {
-
-    case object AmericanExpress extends CreditCardType("AmericanExpress")
-    case object Discover extends CreditCardType("Discover")
-    case object MasterCard extends CreditCardType("MasterCard")
-    case object Visa extends CreditCardType("Visa")
-
-  }
-
-  implicit val writes = new Writes[CreateStripePaymentMethod] {
-    def writes(command: CreateStripePaymentMethod) = Json.obj(
-      "AccountId" -> command.accountId.value,
-      "TokenId" -> command.cardId.value,
-      "SecondTokenId" -> command.customerId.value,
-      "CreditCardCountry" -> command.cardCountry.value,
-      "CreditCardNumber" -> command.last4.value,
-      "CreditCardExpirationYear" -> command.expiration.exp_month,
-      "CreditCardExpirationYear" -> command.expiration.exp_year,
-      "CreditCardType" -> command.creditCardType.value,
-      "Type" -> "CreditCardReferenceTransaction"
-    )
-  }
-
-  implicit val reads: Reads[CreatePaymentMethodResult] =
-    (JsPath \ "Id").read[PaymentMethodId].map(CreatePaymentMethodResult.apply _)
-
-  case class CreatePaymentMethodResult(id: PaymentMethodId)
-
-  def createPaymentMethod(request: CreateStripePaymentMethod): WithDepsFailableOp[ZuoraDeps, CreatePaymentMethodResult] =
-    post[CreateStripePaymentMethod, CreatePaymentMethodResult](request, s"object/payment-method")
-
-}
-
-object ZuoraQueryPaymentMethod {
-
-  case class SecondTokenId(value: String) extends AnyVal
-  case class PaymentMethodId(value: String) extends AnyVal
-  case class AccountId(value: String) extends AnyVal
-  case class CreditCardExpirationMonth(value: Int) extends AnyVal
-  case class CreditCardExpirationYear(value: Int) extends AnyVal
-  case class CreditCardMaskNumber(value: String) extends AnyVal
-  case class PaymentMethodFields(
-    Id: PaymentMethodId,
-    AccountId: AccountId
-  )
-  implicit val fPaymentMethodId: Format[PaymentMethodId] =
-    Format[PaymentMethodId](JsPath.read[String].map(PaymentMethodId.apply), Writes { (o: PaymentMethodId) => JsString(o.value) })
-
-  implicit val fAccountId: Format[AccountId] =
-    Format[AccountId](JsPath.read[String].map(AccountId.apply), Writes { (o: AccountId) => JsString(o.value) })
-
-  implicit val QueryRecordR: Format[PaymentMethodFields] = Json.format[PaymentMethodFields]
-
-  case class AccountPaymentMethodIds(accountId: AccountId, paymentMethodIds: NonEmptyList[PaymentMethodId])
-
-  def getPaymentMethodForStripeCustomer(customerId: StripeCustomerId, sourceId: StripeSourceId): WithDepsFailableOp[ZuoraDeps, AccountPaymentMethodIds] = {
-    import com.gu.util.reader.Types._
-    val query =
-      s"""SELECT Id, AccountId
-         | FROM PaymentMethod
-         |  where Type='CreditCardReferenceTransaction' AND PaymentMethodStatus = 'Active' AND TokenId = '${sourceId.value}' AND SecondTokenId = '${customerId.value}'""".stripMargin
-
-    Zuora.query[PaymentMethodFields](Query(query)).run.map(_.flatMap { result =>
-      result.records.groupBy(_.AccountId).toList match {
-        case (account, first :: rest) :: Nil =>
-          \/-(AccountPaymentMethodIds(account, NonEmptyList(first, rest: _*).map(_.Id)))
-        case _ =>
-          logger.warn(s"wrong number of accounts using the customer token: $result")
-          -\/(ApiGatewayResponse.internalServerError("could not find correct account for stripe details"))
-      }
-    }).toEitherT
-
-  }
-
-}
